Conflicts:
	source/Autowiring/CoreContext.cpp
	source/Autowiring/CoreContext.h
	source/Autowiring/CoreThreadTest.cpp
	source/Autowiring/ExceptionFilterTest.cpp
	source/Autowiring/JunctionBox.cpp
	source/Autowiring/JunctionBox.h

+++ CoreThreadTest.cpp
@@ -169,7 +169,7 @@ TEST_F(CoreThreadTest, VerifyDispatchQueueShutdown) {
   }
   catch (...) {}
 
-  ASSERT_EQ(listener->GetDispatchQueueLength(), static_cast<size_t>(0));
+  ASSERT_EQ(listener->GetDispatchQueueLength(), 0UL);
 }
 
 TEST_F(CoreThreadTest, VerifyNoLeakOnExecptions) {
